+++ packages/@valapi/lib/src/utils/ValEncryption.ts
@@ -1,4 +1,5 @@
 import { Buffer } from "node:buffer";
+import { randomBytes } from "node:crypto";
 
 export class ValEncryption {
     public static encrypt(text: string): string {
@@ -15,5 +16,9 @@ export class ValEncryption {
 
     public static decryptJson<T = any>(object: string): T {
         return JSON.parse(this.decrypt(object));
+    }
+
+    public static randomString(size: number): string {
+        return randomBytes(size).toString("base64url");
     }
 }